moved LinearClassifier in LinearModel.h in compliance with KernelExpansion
