+++ include/shark/Models/LinearModel.h
@@ -34,6 +34,7 @@
 
 #include <shark/Models/AbstractModel.h>
 #include <shark/Models/NeuronLayers.h>
+#include <shark/Models/Classifier.h>
 namespace shark {
 
 
@@ -315,6 +316,26 @@ public:
 	}
 };
 
+/*! \brief Basic linear classifier.
+ *
+ *  The LinearClassifier class is a multi class classifier model
+ *  suited for linear discriminant analysis. For c classes
+ *  \f$ 0, \dots, c-1 \f$  the model computes
+ *   
+ *  \f$ \arg \max_i w_i^T x + b_i \f$
+ *  
+ *  Thus is it a linear model with arg max computation.
+ *  The internal linear model can be queried using decisionFunction().
+ */ 
+template<class VectorType = RealVector>
+class LinearClassifier : public Classifier<LinearModel<VectorType> >
+{
+public:
+	LinearClassifier(){}
+
+	std::string name() const
+	{ return "LinearClassifier"; }
+};
 
 }
 #endif
